Moves log creation to service
Moves the log creation function to the batch render service.
This allows for better code organization and reusability.
Also fixes a bug where the log file was not created if it
did not exist, now creates an empty log with an error message
in this case.
Also ensures to handle encoding errors when reading the log file.